webanalytics-backend/backend-requirements.txt
absl-py==2.1.0
amqp==5.2.0
anyjson==0.3.3
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asgiref==3.7.2
astunparse==1.6.3
async-timeout==4.0.3
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
billiard==4.2.0
bokeh==3.3.4
cachetools==5.3.2
celery==5.3.6
certifi==2024.2.2
cffi==1.16.0
channels==3.0.4
channels-redis==4.2.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
constantly==23.10.4
contourpy==1.2.0
cryptography==42.0.1
cycler==0.12.1
daphne==3.0.2
Django==4.2.9
django-appconf==1.0.6
django-cors-headers==4.3.1
django-cryptography==1.1
django-redis==5.4.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
flatbuffers==23.5.26
fonttools==4.49.0
gast==0.5.4
google-auth==2.27.0
google-auth-oauthlib==1.2.0
google-pasta==0.2.0
greenlet==3.0.3
grpcio==1.60.1
h5py==3.10.0
hyperlink==21.0.0
idna==3.6
incremental==22.10.0
Jinja2==3.1.3
joblib==1.3.2
kiwisolver==1.4.5
kombu==5.3.5
libclang==16.0.6
loguru==0.7.2
Markdown==3.5.2
MarkupSafe==2.1.5
matplotlib==3.8.3
ml-dtypes==0.2.0
msgpack==1.0.7
mysqlclient==2.2.3
numpy==1.26.3
oauthlib==3.2.2
opt-einsum==3.3.0
oracledb==2.0.1
packaging==23.2
pandas==2.2.0
pillow==10.2.0
prompt-toolkit==3.0.43
protobuf==4.25.2
psycopg2==2.9.9
psycopg2-binary==2.9.9
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
PyJWT==2.8.0
pyOpenSSL==24.0.0
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2023.3.post1
PyYAML==6.0.1
redis==5.0.1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==1.4.0
scipy==1.12.0
service-identity==24.1.0
six==1.16.0
SQLAlchemy==2.0.25
sqlparse==0.4.4
tenacity==8.2.3
termcolor==2.4.0
threadpoolctl==3.2.0
tornado==6.4
Twisted==23.10.0
txaio==23.1.1
typing_extensions==4.9.0
tzdata==2023.4
urllib3==2.2.0
vine==5.1.0
wcwidth==0.2.13
Werkzeug==3.0.1
wrapt==1.14.1
xyzservices==2023.10.1
zope.interface==6.1
